Refactor: extracted DirectoryManager class
